BIC-187 # fix post-restructure issues breaking Android 4.3, iOS 7, IE 10-11

+++ src/bic/model-application.js
@@ -188,6 +188,10 @@ define(function (require) {
                 }
               ),
               function (model) {
+                if (!app.hasStorage()) {
+                  app.interactions.remove(model);
+                  return Promise.resolve();
+                }
                 return new Promise(function (resolve, reject) {
                   model.destroy({
                     success: resolve,
@@ -202,6 +206,9 @@ define(function (require) {
           function () {
             app.forms.whenUpdated();
             app.retrieveDataSuitcasesForInteractions();
+            if (!app.hasStorage()) {
+              return Promise.resolve();
+            }
             return app.interactions.save();
           }
         );
